[API] Split /v1 and /api/v1 controllers
